pyproject.toml
[tool.poetry]
name = "image_manupulation"
version = "0.1.0"
description = ""
authors = ["Daviey52 <davidkariska@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
